Added DPPt Trainer Front Pics, added support for u32 pals for trainer pics, commented out mining_minigame from make file due to make error

+++ src/battle_controller_player_partner.c
@@ -390,7 +390,7 @@ static void PlayerPartnerHandleHealthBarUpdate(u32 battler)
 
 static void PlayerPartnerHandleIntroTrainerBallThrow(u32 battler)
 {
-    const u16 *trainerPal;
+    const u32 *trainerPal;
     enum DifficultyLevel difficulty = GetBattlePartnerDifficultyLevel(gPartnerTrainerId);
 
     if (gPartnerTrainerId > TRAINER_PARTNER(PARTNER_NONE))
